added decompilation ratio tests for java
fixed test

renamed mismatching file names for statistics tests

added decompilation ratio tests for java
