+++ daas/daas_app/tests/resources/java_pack.zip
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:53389cce6b83f338a67e431786ba6efaf122cf08b46148e7d7237df6b2ed105c
+size 85191789
