+++ daas/daas_app/tests/test_utils/resource_directories.py
@@ -4,5 +4,6 @@ TEXT_SAMPLE = '/test_resources/text.txt'
 FLASH_SAMPLE_01 = '/test_resources/995bb44df3d6b31d9422ddb9f3f78b7b.flash'
 FLASH_SAMPLE_02 = '/test_resources/eb19009c086845d0408c52d495187380c5762b8c.flash'
 FLASH_ZIPPED_PACK = '/test_resources/flash_pack.zip'
+JAVA_PACK = '/test_resources/java_pack.zip'
 ZIP = '/test_resources/zip.zip'
 ZIP_PROTECTED = '/test_resources/zip_protected_password_is_ASDF1234.zip'
